pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "haste-eeg"
version = "0.1.0"
description = "HASTE: Hybrid-Attention Streaming Topology-aware Estimator for EEG fatigue monitoring"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Qiyu Niu"},
    {name = "Yi Xie"},
    {name = "Siao Liu"},
    {name = "Shouyan Wang"},
    {name = "Yingnan Nie"},
]
keywords = ["eeg", "fatigue", "streaming", "graph-attention", "brain-computer-interface"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=1.13",
    "numpy>=1.21",
    "scipy>=1.7",
    "pyyaml>=6.0",
    "tqdm>=4.60",
    "mne>=1.0",
    "tensorboard>=2.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
haste-train = "train:main"

[tool.setuptools.packages.find]
include = ["models*", "data*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
